Cast data as np.float64 in biweight functions

+++ astropy/stats/biweight.py
@@ -90,7 +90,7 @@ def biweight_location(data, c=6.0, M=None, axis=None):
     -0.0175741540445
     """
 
-    data = np.asanyarray(data)
+    data = np.asanyarray(data).astype(np.float64)
 
     if M is None:
         M = np.median(data, axis=axis)
@@ -324,7 +324,7 @@ def biweight_midvariance(data, c=9.0, M=None, axis=None,
     0.97362869104
     """
 
-    data = np.asanyarray(data)
+    data = np.asanyarray(data).astype(np.float64)
 
     if M is None:
         M = np.median(data, axis=axis)
@@ -510,7 +510,7 @@ def biweight_midcovariance(data, c=9.0, M=None, modify_sample_size=False):
     [ 0.90820237  3.13091961]
     """
 
-    data = np.asanyarray(data)
+    data = np.asanyarray(data).astype(np.float64)
 
     # ensure data is 2D
     if data.ndim == 1:
